:tada: 初步兼容sqlserver

+++ doc/sql/mysql/bladex-sword-mysql.sql
@@ -689,7 +689,7 @@ CREATE TABLE `blade_user`  (
   `email` varchar(45) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL COMMENT '邮箱',
   `phone` varchar(45) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL COMMENT '手机',
   `birthday` datetime(0) NULL DEFAULT NULL COMMENT '生日',
-  `sex` smallint(6) NULL DEFAULT NULL COMMENT '性别',
+  `sex` int(2) NULL DEFAULT NULL COMMENT '性别',
   `role_id` varchar(1000) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL COMMENT '角色id',
   `dept_id` varchar(1000) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL COMMENT '部门id',
   `post_id` varchar(1000) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL COMMENT '岗位id',
